removed unused files
