--- lib/screens/profile.dart
@@ -1,19 +1,0 @@
-// import 'package:flutter/material.dart';
-// import '../widgets/wip_screen.dart';
-
-// class ProfilePage extends StatelessWidget {
-//   const ProfilePage({super.key});
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return Scaffold(
-//       appBar: AppBar(),
-//       body: const Center(
-//         child: Padding(
-//           padding: EdgeInsets.all(8),
-//           child: UnderConstruction(),
-//         ),
-//       ),
-//     );
-//   }
-// }
