--- lib/widgets/wip_screen.dart
@@ -1,23 +1,0 @@
-// import 'package:flutter/material.dart';
-// import "package:lottie/lottie.dart";
-
-// class UnderConstruction extends StatelessWidget {
-//   const UnderConstruction({super.key});
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return Column(
-//       mainAxisAlignment: MainAxisAlignment.center,
-//       children: [
-//         const Text("This page is currently under contruction",
-//             style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24)),
-//         Lottie.asset(
-//           'assets/construction.json',
-//           width: 300,
-//           height: 300,
-//           fit: BoxFit.fill,
-//         ),
-//       ],
-//     );
-//   }
-// }
